Refactor: Renamed controller
